% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{assert}
\usage{
assert(expr, error)
}
\arguments{
\item{expr}{Check Expression.}

\item{error}{A character string containing an error message for the assertion when it fails.}
}
\description{
The purpose of this function is to raise error messages in the case of a function receiving an invalid input.
}
\examples{
delta <- 0.1
assert(delta > 0 & delta < 1, "delta must be between 0 and 1")

}
\author{
Christian Voldby, Data Science student at Aarhus University
}
