% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_data_manipulation.R
\name{cover_data_manipulation}
\alias{cover_data_manipulation}
\title{cover_data_manipulation}
\usage{
cover_data_manipulation(data, site_id, plant1_id, plant2_id, min_number_of_plots = 30)
}
\arguments{
\item{data}{The "alledata-cover.csv" data frame.}

\item{site_id}{The side id for the side of interest.}

\item{plant1_id}{The plant id must begin with a "X" followed by the id number.}

\item{plant2_id}{The plant id must begin with a "X" followed by the id number.}

\item{min_number_of_plots}{Positive integer.}
}
\value{
A data frame containing the relevant cover data for the two chosen plants.
}
\description{
Data manipulation:
The purpose of this function is to easily obtain the data of interest from the cover data.
The dataset is the alledata-cover . csv provided by the instructor Christian Damsgaard.
Therefore the function must include the column 'site' from the dataset.
}
\author{
Christian Voldby, Data Science student at Aarhus University
}
