% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_cov.R
\name{expected_cov}
\alias{expected_cov}
\title{The Expected Covariance}
\usage{

}
\value{
single numeric value
}
\description{
The purpose of this function is to calculate expected covariance at a site.
}
\examples{


}
\author{
Christian Voldby, Data Science student at Aarhus University
}
