% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraplot_corr.R
\name{intraplot_corr}
\alias{intraplot_corr}
\title{intra-plot correlation}
\usage{
intraplot_corr(data, site_id, species_id, year)
}
\arguments{
\item{data}{the cover data.}

\item{site_id}{the site of interest.}

\item{species_id}{identifier for a species in the cover data.}

\item{year}{the year of interest.}
}
\value{
single numeric value.
}
\description{
The purpose of this function is to estimate intra-plot correlation parameter that measures the within-site spatial aggregation of the plant species as the correlation between the outcomes of successive Bernoulli trials.
}
\author{
Christian Voldby, Data Science student at Aarhus University.
}
