% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_cover.R
\name{mean_cover}
\alias{mean_cover}
\title{mean cover}
\usage{
mean_cover(data , site_id , species_id , year)
}
\arguments{
\item{data}{The 'alledata-cover . csv' data.frame.}

\item{site_id}{The site of interest.}

\item{species_id}{Identifier for a species in the cover data for instance X974.}

\item{year}{The year of interest. To not mix cover data from different years.}
}
\value{
Single numeric value
}
\description{
The purpose is to estimate mean plant cover at a site.
The value must be between 0 and 1.
The dataset is the 'alledata-cover . csv' provided by the instructor Christian Damsgaard.
}
\author{
Christian Voldby, Data Science student at Aarhus University
}
