% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_cov.R
\name{observed_cov}
\alias{observed_cov}
\title{The Observed Covariance}
\usage{
observed_cov(r, s, t, u, N)
}
\arguments{
\item{r}{vector with the number of pins that only touch plant species A in each plot at a site.}

\item{s}{vector with the number of pins that only touch plant species B in each plot at a site.}

\item{t}{vector with the number of pins that touch both plant species A and B in each plot at a site.}

\item{u}{vector with the number of pins that touch neither plant species A or B in each plot at a site.}

\item{N}{number of plots at a site.}
}
\value{
single numeric value
}
\description{
The purpose of this function is to calculate the observed covariance.
The function for this observed covariance is corrected by the number of plots at a site.
}
\examples{
r <- c(1,2,3,4,5)
s <- c(5,4,3,2,1)
t <- c(2,3,2,3,2)
u <- c(8,7,8,7,8)
N <- 5
observed_cov(r, s, t, u, N)

}
\author{
Christian Voldby, Data Science student at Aarhus University
}
