% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_beta.R
\name{p_beta}
\alias{p_beta}
\title{p_beta}
\usage{

}
\arguments{
\item{delta}{The intra-plot corelation parameter}

\item{q}{the mean cover parameter}
}
\value{
single numeric value
}
\description{
The purpose of this function is adjust the beta distribution to the parameters delta and q.
}
\examples{

}
\author{
Christian Voldby, Data Science student at Aarhus University
}
