% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstu.R
\name{rstu}
\alias{rstu}
\title{RSTU}
\usage{
rstu(cover_data)
}
\arguments{
\item{cover_data}{the cover data must only include the columns containing the cover data.}
}
\value{
matrix(r, s, t, u)
}
\description{
The purpose of this function is to easily get the r,s,t,u variables for a site.
The data must first be clean and ready for the process. This could be done by first applying the 'cover_data_manipulation' function.
}
\author{
Christian Voldby, Data Science student at Aarhus University
}
