% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_plant_cover_data.R
\name{simulate_plant_cover_data}
\alias{simulate_plant_cover_data}
\title{simulate_plant_cover_data}
\usage{
simulate_plant_cover_data(N)
}
\arguments{
\item{N}{the number of plots in a site.}
}
\value{
matrix(r, s, t, u)
}
\description{
The purpose of this function is to simulate N plots of plant cover data. Thus returning r, s, t, u.
The simulation has mainly been used in order to have some easy data to test the other functions with.
}
\examples{
N <- 45
simulate_plant_cover_data(N)

}
\author{
Christian Voldby, Data Science student at Aarhus University
}
