% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_cover.R
\name{mean_cover}
\alias{mean_cover}
\title{mean cover}
\usage{

}
\arguments{
\item{data}{the cover data.frame}

\item{site_id}{the site of interest}

\item{species_id}{identifier for a species in the cover data.}
}
\value{
single numeric value
}
\description{
The purpose is to estimate mean plant cover at a site.
}
\examples{

}
\author{
Christian Voldby, Data Science student at Aarhus University
}
