% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya.R
\name{polya}
\alias{polya}
\title{Polya-Eggenberger distribution of pin-point hits for a single species.}
\usage{

}
\arguments{
\item{x}{vector of quantiles}

\item{delta}{parameter for the intra-plot correlation (\code{0 < delta < 1})}

\item{q}{parameter for the mean cover (\code{0 < delta < 1})}

\item{n}{number of pins in a plot. The default is 16 pins.}
}
\value{
single numeric value
}
\description{
The purpose is the be a distribution of pin-point hits for a single species.
}
\examples{
x <- 5
delta <- 3
q <- 0.5
polya(x, delta, q)

}
\author{
Christian Voldby, Data Science student at Aarhus University
}
