% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopinpoint.R
\name{twopinpoint}
\alias{twopinpoint}
\title{The distribution of pin-point hits for two species}
\usage{
twopinpoint(r, s, t, u, N)
}
\description{
The purpose of the function is to act as a distribution of pin-point hits for two species.
}
\details{
Four stochastic variables may be defined for each plot:
- r := the number of pins that only touch plant species A.
- s := the number of pins that only touch plant species B.
- t := the number of pins that touch both plant species A and B.
- u := the number of pint that touch neither species A or B.
}
